Add script for data clean up
The basic functionality of the script is to read the training and test
datasets - in csv format - from the `data` directory present in the
project workspace. The pre-requisite of running this script is to ensure
that the respective files are present in the directory.

The process begins by reading the files' content into dataframes. The
rows of data are then processed, with the first step being dropping the
empty rows.

The emojis contained in the text are then separated by spaces to better
identify them.

The common stopwords in English language are then cleaned up from the
text to avoid data skewing.

Finally, unwanted symbols are removed from the text - such as URLs, user
handles, punctuations, line breaks, etc.

The duplicate rows are also removed to end the process.

The cleaned data are then stored in respective files in the `clean_data`
directory, which is created first in the project workspace, if it does
not exist.