Add script for training model for Task A
The first task of the project was to identify if the provided text is
sarcastic or not, by using LSTM recurrent neural network algorithm.

The process begins by reading the cleaned-up training and test datasets
from the `clean_data` directory. The training data is split into two
parts - one for training and the other for validation during training -
into 70-30 splits.

The training split is then used to form a TextVectorization layer,
forming text vectors from the texts provided.

A sequential pipefile is then formed to create the model, which contains
the vectorizer layer first, followed by an Embedding layer that creates
text embeddings from the word vectors. The output dimension of the
Embedding layer is 64. These embeddings are then passed to a
bidirectional LSTM layer consisting of 64 units, with 0.5 dropout, and
L2 regularizers. The output of the LSTM layers is passed through a Dense
layer to obtain the final binary output.

The model was also equipped to use Adam optimizer. For loss calculation,
binary cross-entropy was used, while F1 score was calculated to measure
the performance of the model.

The model was then fit against the training data and validated with the
validation data across 30% of the training data. The model was trained
for 10 epochs.

Afterwards, the validation data was used to evaluate the model, and the
loss and F1 scores were measured and plotted. Finally, the test dataset
was put into test, obtaining the loss and performance metrics generated by
the model.