Add script for training model for task B
The second task of the project was similar, but this time, the text had
to be classified across six levels of irony speech.

A similar approach was followed to split the training, validation and
test data, as well as to define the model. The parameters were changed
due to the size of the training data, however. This time around, the
LSTM layer contained 32 units. The final dense layers contained 6 units
in order to output predictions for the all 6 class labels.

The learning rate of the Adam optimizer was well decreased for the model
to better fit the training data. For measuring performance of the model
that outputs 6 class labels, macro F1 Score was measured.

The same procedure was followed to obtain the result of the task - the
model was trained with the training data, with a portion of it being
used for validation. The true validation dataset was then used to
evaluate the model, before the actual test dataset was used to find the
error and measure the performance (i.e. F1 Score) of the model.