Add Jupyter Notebook deliverables
The scripts are converted to Jupyter Notebooks, in order for the code to
run anywhere and everywhere - in any Conda environment, Pycharm,
Datalore, Google Collab, etc.

A "how-to.txt" file describes how to run these notebooks, including the
means of installing the dependencies, finding the datasets, and the
pre-requisites required to run the files.